+++ src/screens/Payment/modules/CardsModuleSlidable.tsx
@@ -1,7 +1,7 @@
-import React, { useState } from 'react';
+import React from 'react';
 import { Box, HStack, Text, VStack } from 'native-base';
 import { Pressable, StyleSheet } from 'react-native';
-import { CreditCard, EmptyCreditCard, PressableIcon, ActionButton, FormInput } from '../../../components/shared';
+import { CreditCard, EmptyCreditCard } from '../../../components/shared';
 import { MaterialIcons } from '@expo/vector-icons';
 import { CreditCard as CreditCardType } from '../../../types/CreditCard';
 import CategoryMap from '../../../components/shared/CreditCard/utils/category-map';
